Fix signed/unsigned comparison bug in NeighborLists
- Bump to version 1.7.4
- Issue arose when using size_t for number_of_neighbors and
  number_of_neighbors_lists
- Now covered by NeighborSearchTest radius search
- Also switched team scratch size calculation to size_t

+++ examples/NeighborSearchTest.cpp
@@ -171,9 +171,9 @@ bool all_passed = true;
             // CreatePointCloudSearch constructs an object of type PointCloudSearch, but deduces the templates for you
             auto point_cloud_search(CreatePointCloudSearch(source_coords, dimension));
 
-            Kokkos::View<int*, Kokkos::DefaultHostExecutionSpace> neighbor_lists("neighbor lists", 
+            Kokkos::View<size_t*, Kokkos::DefaultHostExecutionSpace> neighbor_lists("neighbor lists", 
                     number_target_coords); // first column is # of neighbors
-            Kokkos::View<int*, Kokkos::DefaultHostExecutionSpace> number_neighbors_list("number of neighbors list", 
+            Kokkos::View<size_t*, Kokkos::DefaultHostExecutionSpace> number_neighbors_list("number of neighbors list", 
                     number_target_coords); // first column is # of neighbors
             
             // each target site has a window size
